"add hero integrated with API"

+++ src/app/heroes/heroes.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
-import { Hero } from 'src/app/core/models/hero.model';
+import { Hero, HeroPagination } from 'src/app/core/models/hero.model';
 import { HeroService } from 'src/app/services/hero.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-heroes',
@@ -8,7 +9,7 @@ import { HeroService } from 'src/app/services/hero.service';
   styleUrls: ['./heroes.component.scss'],
 })
 export class HeroesComponent implements OnInit {
-  displayedColumns: string[] = ['id', 'name', 'actions'];
+  displayedColumns: string[] = ['idHero', 'name', 'actions'];
   heroes: Hero[] = [];
 
   constructor(private heroService: HeroService) {}
@@ -18,13 +19,30 @@ export class HeroesComponent implements OnInit {
   }
 
   getHeroes(): void {
-    this.heroService.getHeroes().subscribe((heroes) => (this.heroes = heroes));
+    this.heroService.getHeroes().subscribe((heroes: HeroPagination) => {
+
+      console.log(heroes.items)
+      this.heroes = heroes.items;
+    });
   }
 
-
   deleteHero(id: number): void {
-    this.heroService.deleteHero(id).subscribe(() => {
-      this.heroes = this.heroes.filter((hero) => hero.id !== id);
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.heroService.deleteHero(id).subscribe(() => {
+          // this.heroes = this.heroes.filter((hero) => hero.idHero !== id);
+          Swal.fire('Deleted!', 'Your hero has been deleted.', 'success');
+        });
+      }
+      console.log(result);
     });
   }
 }
